# docs/openapi.yaml
openapi: 3.0.0
info:
  title: DevRank API
  description: A comprehensive platform for developers to showcase projects, earn certifications, manage skills, and build professional networks through code reviews and badges.
  version: 1.0.0
  contact:
    name: DevRank Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.devrank.com
    description: Production server
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users with their skills and profile information
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /projects:
    post:
      tags:
        - Projects
      summary: Create a new project
      description: Create a new project with title, description, technologies, and details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  project:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found

    get:
      tags:
        - Projects
      summary: Get all projects
      description: Retrieve all projects with optional filtering by userId, status, or techStack
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: status
          in: query
          schema:
            type: string
            enum: [seeking-review, reviewed, archived]
          description: Filter by project status
        - name: techStack
          in: query
          schema:
            type: string
          description: Filter by technology
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /reviews:
    post:
      tags:
        - Reviews
      summary: Create a new review
      description: Submit a review for a project as a verified reviewer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  review:
                    $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Only verified reviewers can submit reviews
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already reviewed by this reviewer

    get:
      tags:
        - Reviews
      summary: Get all reviews
      description: Retrieve all reviews with optional filtering by projectId, reviewerId, or status
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
          description: Filter by project ID
        - name: reviewerId
          in: query
          schema:
            type: string
          description: Filter by reviewer ID
        - name: status
          in: query
          schema:
            type: string
            enum: [published, removed]
          description: Filter by review status
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /skills:
    post:
      tags:
        - Skills
      summary: Create a new skill
      description: Add a new skill to the platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: Skill created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  skill:
                    $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Skill already exists

    get:
      tags:
        - Skills
      summary: Get all skills
      description: Retrieve all skills with optional filtering by category or preset
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: preset
          in: query
          schema:
            type: boolean
          description: Filter by preset status
      responses:
        '200':
          description: Skills retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /certifications:
    get:
      tags:
        - Certifications
      summary: Get all certification requests
      description: Retrieve all certification requests (admin only)
      responses:
        '200':
          description: Certification requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  certifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/CertificationRequest'
        '403':
          description: Unauthorized

    /certifications/{certificationId}:
      get:
        tags:
          - Certifications
        summary: Get certification by ID
        description: Retrieve a specific certification by its ID
        parameters:
          - name: certificationId
            in: path
            required: true
            schema:
              type: string
            description: Certification ID (MongoDB ObjectId)
        responses:
          '200':
            description: Certification found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    certification:
                      $ref: '#/components/schemas/CertificationRequest'
          '404':
            $ref: '#/components/responses/NotFound'
...
  /logs:
    get:
      tags:
        - Activity Logs
      summary: Get all activity logs
      description: Retrieve all activity logs with optional filtering by userId, action, entity, entityId, or date range
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: action
          in: query
          schema:
            type: string
          description: Filter by action type
        - name: entity
          in: query
          schema:
            type: string
          description: Filter by entity type
        - name: entityId
          in: query
          schema:
            type: string
          description: Filter by entity ID
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter from start date (ISO 8601)
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter until end date (ISO 8601)
      responses:
        '200':
          description: Activity logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLog'
        '400':
          $ref: '#/components/responses/BadRequest'

    delete:
      tags:
        - Activity Logs
      summary: Delete all logs
      description: Delete all activity logs (use with caution)
      responses:
        '200':
          description: All logs deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedCount:
                    type: number
...
components:
  schemas:
...
    CreateProjectRequest:
      type: object
      required:
        - userId
        - title
        - description
      properties:
        userId:
          type: string
        title:
          type: string
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
...
    CreateReviewRequest:
      type: object
      required:
        - projectId
        - reviewerId
        - overallRating
        - codeQualityScore
        - creativityScore
        - cleanCodeScore
      properties:
        projectId:
          type: string
        reviewerId:
          type: string
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
        codeQualityScore:
          type: integer
          minimum: 1
          maximum: 5
        creativityScore:
          type: integer
          minimum: 1
          maximum: 5
        cleanCodeScore:
          type: integer
          minimum: 1
          maximum: 5
        wouldHire:
          type: boolean
        generalFeedback:
          type: string
          maxLength: 1000
        suggestions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [published, removed]
...
    CreateSkillRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        category:
          type: string
        icon:
          type: string
          format: uri
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [developer, reviewer]
        bio:
          type: string
        githubUrl:
          type: string
        avatarUrl:
          type: string
        isVerifiedReviewer:
          type: boolean
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
        followers:
          type: array
          items:
            type: string
        following:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        repositoryUrl:
          type: string
        demoUrl:
          type: string
        status:
          type: string
          enum: [seeking-review, reviewed, archived]
        aggregateRating:
          type: number
        aggregateCodeQuality:
          type: number
        aggregateCreativity:
          type: number
        aggregateCleanCode:
          type: number
        totalReviews:
          type: integer
        hireVotes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Review:
      type: object
      properties:
        _id:
          type: string
        project:
          $ref: '#/components/schemas/Project'
        reviewer:
          $ref: '#/components/schemas/User'
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
        codeQualityScore:
          type: integer
          minimum: 1
          maximum: 5
        creativityScore:
          type: integer
          minimum: 1
          maximum: 5
        cleanCodeScore:
          type: integer
          minimum: 1
          maximum: 5
        wouldHire:
          type: boolean
        generalFeedback:
          type: string
        suggestions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [published, removed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Skill:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        preset:
          type: boolean
        users:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Badge:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        criteria:
          type: string
        isActive:
          type: boolean
        users:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CertificationRequest:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        certificationName:
          type: string
        issuer:
          type: string
        certificationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, approved, rejected]
        approvalMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        action:
          type: string
        entity:
          type: string
        entityId:
          type: string
        metadata:
          type: object
        timestamp:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        errorCode:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'